CRUD place types

+++ src/app/admin/center/place-type/place-type.page.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { PlacesService } from 'src/app/core/services/places.service';
 import { AlertController } from '@ionic/angular';
 import { PlacesWasteTypes } from 'src/app/core/models/waste-type';
+import { Place } from '../../../core/models/place.model';
 
 @Component({
   selector: 'app-place-type',
@@ -17,6 +18,7 @@ import { PlacesWasteTypes } from 'src/app/core/models/waste-type';
    */
 export class PlaceTypePage implements OnInit {
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   placeTypes: any [];
   placeWasteTypeToDelete: PlacesWasteTypes[];
 
@@ -58,29 +60,53 @@ export class PlaceTypePage implements OnInit {
    * @returns 
    */
   onDeletePlaceType(placeTypeId: string, name: string) {
-    this.alertCtrl.create ({
-      header: 'Mensaje de Confirmación',
-      message: '¿De verdad quieres eliminar el tipo de residuo "'+name + '"?',
-      buttons: [{
-        text: 'Cancelar',
-        role: 'cancel'
-      }, {
-        text: 'Borrar',
-        handler: () => {
-          this.placesService.deletePlaceTypeByID(placeTypeId).then(() => {
-            this.placesService.getAllWasteTypeByPlaceType(placeTypeId).then(data => {
-              this.placeWasteTypeToDelete = data;
-              for (const item of this.placeWasteTypeToDelete){
-                this.placesService.deletePlaceWasteType(item.id);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    let placeType: any[];
+    this.placesService.getPlaceTypeByID(placeTypeId).then(item => {
+      placeType = [{
+        place_type: item.id
+      }];
+      this.placesService.getIDPlacesByPlacesType(placeType).then(data => {
+        let places: Place[] = data;
+        // eslint-disable-next-line @typescript-eslint/no-magic-numbers
+        if(places.length > 0){
+          this.alertCtrl.create ({
+            header: 'Atención',
+            message: 'No es posible eliminar la categoría de desecho "' + name + '" ya que está asociada a diversos centros.',
+            buttons: [{
+              text: 'Aceptar',
+              role: 'cancel'
+            }]
+          }).then(alertEl => {
+            alertEl.present();
+          });
+        }else{
+          this.alertCtrl.create ({
+            header: 'Mensaje de Confirmación',
+            message: '¿De verdad quieres eliminar el tipo de residuo "'+ name + '"?',
+            buttons: [{
+              text: 'Cancelar',
+              role: 'cancel'
+            }, {
+              text: 'Borrar',
+              handler: () => {
+                this.placesService.deletePlaceTypeByID(placeTypeId).then(() => {
+                  this.placesService.getAllWasteTypeByPlaceType(placeTypeId).then(data => {
+                    this.placeWasteTypeToDelete = data;
+                    for (const item of this.placeWasteTypeToDelete){
+                      this.placesService.deletePlaceWasteType(item.id);
+                    }
+                  });
+                  this.placesService.allPlaceTypes().then( data => { this.placeTypes = data; });
+                })
+                    .catch(() => {});
               }
-            });
-            this.placesService.allPlaceTypes().then( data => { this.placeTypes = data; });
-          })
-              .catch(() => {});
+            }]
+          }).then(alertEl => {
+            alertEl.present();
+          });
         }
-      }]
-    }).then(alertEl => {
-      alertEl.present();
-    });
+      });
+    });   
   }
 }
